refactor: extract group menu overlay (#357)
Thirteenth slice. Largest single overlay, multi-screen.

draw_group_menu (~280 lines) handles six sub-screens via a match on
group_menu.state:
- Menu: per-group action list
- Members: scrollable member list
- AddMember / RemoveMember: type-to-filter contact pickers
- Rename / Create: text-input popups
- LeaveConfirm: y/n prompt

GroupMenuState moves with it; ui/mod.rs no longer imports it.

ui/mod.rs: 2747 -> 2466 lines.

This was the last overlay still in mod.rs. Going forward, ui/overlays/
contains all 19 overlay submodules; the remaining work in mod.rs is
chat-area rendering (draw_messages/welcome/chat_area), the autocomplete
popup (composer-adjacent), and link/image post-render helpers.
